display: theme holds styling, charting holds charts, formatting holds units
